llm_engine/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
openai==1.18.0
llama-cpp-python==0.2.56
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1